% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_stats.R
\name{omni_cohen_d}
\alias{omni_cohen_d}
\title{Cohen's d (pooled SD) and Welch's t-test.}
\usage{
omni_cohen_d(x, y)
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
Cohen's d (pooled SD).
}
\description{
Cohen's d (pooled SD) and Welch's t-test.
}
