% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_stats.R
\name{omni_eb_shrink}
\alias{omni_eb_shrink}
\title{Empirical-Bayes (method-of-moments) shrinkage of estimates toward the grand mean.}
\usage{
omni_eb_shrink(beta, se)
}
\arguments{
\item{beta}{Numeric estimates.}

\item{se}{Numeric standard errors.}
}
\value{
Shrunk estimates (with attribute \code{tau2}).
}
\description{
Empirical-Bayes (method-of-moments) shrinkage of estimates toward the grand mean.
}
