% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_spatial.R
\name{omni_geary}
\alias{omni_geary}
\title{Global Geary's C. E[C] = 1.}
\usage{
omni_geary(x, W, n_perm = 999L, seed = 0L)
}
\arguments{
\item{x}{Numeric variable.}

\item{W}{Spatial weight matrix.}

\item{n_perm}{Number of permutations.}

\item{seed}{RNG seed.}
}
\value{
A list with \code{C}, \code{EC}, \code{p_perm}.
}
\description{
Global Geary's C. E[C] = 1.
}
