% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_spatial.R
\name{omni_moran}
\alias{omni_moran}
\title{Global Moran's I with a permutation null. E[I] = -1/(n-1).}
\usage{
omni_moran(x, W, n_perm = 999L, seed = 0L)
}
\arguments{
\item{x}{Numeric variable.}

\item{W}{Spatial weight matrix.}

\item{n_perm}{Number of permutations.}

\item{seed}{RNG seed.}
}
\value{
A list with \code{I}, \code{EI}, \code{Z}, \code{p_perm}, \code{null_sd}.
}
\description{
Global Moran's I with a permutation null. E[I] = -1/(n-1).
}
