% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_stats.R
\name{omni_mwu}
\alias{omni_mwu}
\title{Mann-Whitney U p-value (normal approximation with tie correction).}
\usage{
omni_mwu(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x, y}{Numeric vectors to compare.}

\item{alternative}{One of "two.sided", "greater", "less".}
}
\value{
A list with \code{statistic} (U), \code{z} and \code{p.value}.
}
\description{
Mann-Whitney U p-value (normal approximation with tie correction).
}
