% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{omni_rnames}
\alias{omni_rnames}
\alias{omni_cnames}
\title{Row / column names with a safe fallback.}
\usage{
omni_rnames(X)

omni_cnames(X)
}
\arguments{
\item{X}{A matrix-like object.}
}
\value{
A character vector of names (or indices if unnamed).
}
\description{
Row / column names with a safe fallback.
}
