% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_stats.R
\name{omni_perm_p}
\alias{omni_perm_p}
\title{Empirical permutation p-value from a null distribution and an observed statistic.}
\usage{
omni_perm_p(obs, null, alternative = c("greater", "two.sided"))
}
\arguments{
\item{obs}{Observed statistic.}

\item{null}{Numeric null distribution.}

\item{alternative}{One of "greater", "two.sided".}
}
\value{
The empirical p-value.
}
\description{
Empirical permutation p-value from a null distribution and an observed statistic.
}
