% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_program.R
\name{omni_ucell_signed}
\alias{omni_ucell_signed}
\title{Signed UCell programme score = UCell(up) - UCell(down).}
\usage{
omni_ucell_signed(X, up, down, ...)
}
\arguments{
\item{X}{cells x genes matrix.}

\item{up, down}{Character vectors of up- and down-programme genes.}

\item{...}{Passed to \code{omni_ucell}.}
}
\value{
A numeric vector (UCell(up) - UCell(down)).
}
\description{
Signed UCell programme score = UCell(up) - UCell(down).
}
