% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_stats.R
\name{omni_welch}
\alias{omni_welch}
\title{Welch's t-test (unequal variances).}
\usage{
omni_welch(x, y)
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
The \code{htest} object from \code{stats::t.test}.
}
\description{
Welch's t-test (unequal variances).
}
